add_rogin

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,18 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+    
         
-        return view('tasks.index',[
+        $user = \Auth::user();
+        $tasks = $user->tasks;
+        
+        $data = [
+            'user'  => $user,
             'tasks' => $tasks,
-        ]);
+            ];
+    
+      
+        return view('tasks.index', $data);
     }
 
     /**
@@ -49,11 +56,11 @@ class TasksController extends Controller
             'status'  => 'required|max:10',
         ]);
         
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-        
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status'  => $request->status,
+        ]);
+    
         return redirect('/');
     }
 
@@ -66,10 +73,14 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::find($id);
-        
-        return view('tasks.show',[
-            'task' => $task,
-            ]);
+        if (\Auth::id() === $task->user_id) {
+            
+            return view('tasks.show',[
+                'task' => $task,
+                ]);
+        } else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -82,9 +93,13 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
         
-        return view('tasks.edit',[
-            'task' => $task,
-            ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit',[
+                'task' => $task,
+                ]);
+        } else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -117,9 +132,13 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
+        
         $task = Task::find($id);
-        $task->delete();
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
         
         return redirect('/');
+        
     }
 }
